Organize folders in dev/app (closes #521) (#524)
