+++ dev/app/spec/lib/routing/Router_spec.js
@@ -1,8 +1,8 @@
 // Unit tests: exercises the Router class in isolation inside a minimal Express wrapper.
 import request from 'supertest';
-import { ALL_CATEGORIES, BOOKS_CATEGORY, HOBBIT_ITEM } from '../support/fixtures/expectedResponses.js';
-import { testData as data } from '../support/fixtures/testData.js';
-import { buildRouterApp } from '../support/utils/AppFactory.js';
+import { ALL_CATEGORIES, BOOKS_CATEGORY, HOBBIT_ITEM } from '../../support/fixtures/expectedResponses.js';
+import { testData as data } from '../../support/fixtures/testData.js';
+import { buildRouterApp } from '../../support/utils/AppFactory.js';
 
 describe('Router', () => {
   const app = buildRouterApp(data);
